[Laura] updated to get rid of merge conflict due to quantity data type (from int to double)

+++ Android/app/src/main/java/com/example/callouskitchenandroid/AddFoodActivity.kt
@@ -31,9 +31,9 @@ class AddFoodActivity : AppCompatActivity() {
             {
                 // todo: add the food with the service
                 val quantityString = txtFoodQuantity.text.toString()
-                var quantity: Int = 1
+                var quantity: Double = 1.0
                 if (!quantityString.isNullOrEmpty())
-                    quantity = quantityString.toInt()
+                    quantity = quantityString.toDouble()
 
                 val food = Food(1, foodName, quantity)
